refactor(command): move checkVersion method to separate file (#623)
